02-entra-saml-sso/requirements.txt
blinker==1.9.0
click==8.3.3
Flask==3.1.3
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.1.0
MarkupSafe==3.0.3
python-dotenv==1.2.2
python3-saml==1.16.0
Werkzeug==3.1.8
xmlsec==1.3.17
